requirements.txt
Flask==2.2.2
openai==0.27.8
PyPDF2==3.0.1
python-dotenv===1.0.0